fix: build-time config missing properties
