+++ quarkus-integration/quarkus/deployment/src/test/java/ai/timefold/solver/quarkus/TimefoldProcessorMultipleSolverPropertiesTest.java
@@ -1,12 +1,8 @@
 package ai.timefold.solver.quarkus;
 
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNotSame;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import jakarta.inject.Inject;
-import jakarta.inject.Named;
-
-import ai.timefold.solver.core.api.solver.SolverManager;
+import ai.timefold.solver.quarkus.rest.TestdataQuarkusSolutionConfigResource;
 import ai.timefold.solver.quarkus.testdata.normal.constraints.TestdataQuarkusConstraintProvider;
 import ai.timefold.solver.quarkus.testdata.normal.domain.TestdataQuarkusEntity;
 import ai.timefold.solver.quarkus.testdata.normal.domain.TestdataQuarkusSolution;
@@ -17,36 +13,21 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
 import io.quarkus.test.QuarkusUnitTest;
+import io.restassured.RestAssured;
 
 class TimefoldProcessorMultipleSolverPropertiesTest {
 
     @RegisterExtension
     static final QuarkusUnitTest config = new QuarkusUnitTest()
-            .overrideConfigKey("quarkus.timefold.solver.\"solver1\".environment-mode", "FULL_ASSERT")
-            .overrideConfigKey("quarkus.timefold.solver.\"solver1\".daemon", "true")
-            .overrideConfigKey("quarkus.timefold.solver.\"solver1\".domain-access-type", "REFLECTION")
-            .overrideConfigKey("quarkus.timefold.solver.\"solver1\".termination.spent-limit", "4h")
-            .overrideConfigKey("quarkus.timefold.solver.\"solver1\".termination.unimproved-spent-limit", "5h")
-            .overrideConfigKey("quarkus.timefold.solver.\"solver1\".termination.best-score-limit", "0")
-            .overrideConfigKey("quarkus.timefold.solver.\"solver2\".daemon", "false")
-            .overrideConfigKey("quarkus.timefold.solver.\"solver2\".domain-access-type", "REFLECTION")
-            .overrideConfigKey("quarkus.timefold.solver.\"solver2\".termination.spent-limit", "1h")
+            .overrideConfigKey("quarkus.timefold.solver.\"solver1\".termination.spent-limit", "4s")
+            .overrideConfigKey("quarkus.timefold.solver.\"solver2\".termination.spent-limit", "2s")
             .setArchiveProducer(() -> ShrinkWrap.create(JavaArchive.class)
                     .addClasses(TestdataQuarkusEntity.class, TestdataQuarkusSolution.class,
-                            TestdataQuarkusConstraintProvider.class));
-
-    @Inject
-    @Named("solver1")
-    SolverManager<TestdataQuarkusSolution, String> solver1;
-
-    @Inject
-    @Named("solver2")
-    SolverManager<TestdataQuarkusSolution, String> solver2;
+                            TestdataQuarkusConstraintProvider.class, TestdataQuarkusSolutionConfigResource.class));
 
     @Test
     void solverProperties() {
-        assertNotNull(solver1);
-        assertNotNull(solver2);
-        assertNotSame(solver1, solver2);
+        String resp = RestAssured.get("/solver-config/seconds-spent-limit").asString();
+        assertEquals("secondsSpentLimit=0.12;secondsSpentLimit=0.25", resp);
     }
 }
